Fix training issues: dtype, OOM, and BatchNorm errors
- Fix label dtype: ensure labels are cast to int64/long for CrossEntropyLoss
- Reduce batch_size from 4 to 2 to prevent OOM on GPUs with limited VRAM
- Add drop_last=True to train DataLoader to prevent BatchNorm failure on single-sample batches
- Fix labels_dir path in DeepLabV3 configs (data/labels -> data/masks/npy)

+++ src/training/trainer.py
@@ -162,7 +162,7 @@ class Trainer:
 
         for batch_idx, (images, labels) in enumerate(pbar):
             images = images.to(self.device)
-            labels = labels.to(self.device)
+            labels = labels.to(self.device).long()
 
             self.optimizer.zero_grad()
             outputs = self.model(images)
@@ -194,7 +194,7 @@ class Trainer:
         with torch.no_grad():
             for images, labels in pbar:
                 images = images.to(self.device)
-                labels = labels.to(self.device)
+                labels = labels.to(self.device).long()
 
                 outputs = self.model(images)
                 loss = self.loss_fn(outputs, labels)
